Use Swift-Language-Mods for eachWithObject (#2)
Note that with Swift-Language-Mods version 1.0.0, this hardwires the XCTest dependency into any executables we make (it uses the Spec package).  The problem with that is that command-line executables will need to be dynamically lined against XCTest even though they run no tests.  This will only be a problem until Swift Packages support separate test dependencies.